Image-Based/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
tensorflow>=2.8.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
Pillow>=8.0.0
pillow-heif>=0.8.0
ultralytics>=8.0.0
tqdm>=4.65.0
